Перенесена логика запроса из WebViewViewController в WebViewPresenter
