+++ ImageFeed/Auth/WebViewPresenter.swift
@@ -1,0 +1,40 @@
+import Foundation
+
+public protocol WebViewPresenterProtocol {
+    var view: WebViewViewControllerProtocol? { get set }
+    func viewDidLoad()
+}
+
+final class WebViewPresenter: WebViewPresenterProtocol {
+    weak var view: WebViewViewControllerProtocol?
+    
+    enum WebViewConstants {
+        static let unsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
+    }
+    
+    func viewDidLoad() {
+        loadAuthView()
+    }
+    
+    private func loadAuthView() {
+        guard var urlComponents = URLComponents(string: WebViewConstants.unsplashAuthorizeURLString) else {
+            print("Error creating URLComponents")
+            return
+        }
+        
+        urlComponents.queryItems = [
+            URLQueryItem(name: "client_id", value: Constants.accessKey),
+            URLQueryItem(name: "redirect_uri", value: Constants.redirectURI),
+            URLQueryItem(name: "response_type", value: "code"),
+            URLQueryItem(name: "scope", value: Constants.accessScope)
+        ]
+        
+        guard let url = urlComponents.url else {
+            print("URLComponents is equal to nil")
+            return
+        }
+        
+        let request = URLRequest(url: url)
+        view?.load(request: request)
+    }
+}
